refactor: extract ExportReceiptStoring.swift from ExportReceipt.swift
Splits the `ExportReceiptStoring` protocol out of ExportReceipt.swift
so that file holds only the receipt value type.

Byte-identical move of lines 33-51. Both files import Foundation.

ExportReceipt.swift: 51 → 31 lines. Tests: 667.
